# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Firakef1_settle_backend_internal_requests_dto.AddCommentDTO:
    properties:
      content:
        type: string
    type: object
  github_com_Firakef1_settle_backend_internal_requests_dto.ApprovalResponse:
    properties:
      status:
        description: Simple stub for approval domain interactions if needed later
        type: string
    type: object
  github_com_Firakef1_settle_backend_internal_requests_dto.CommentResponse:
    properties:
      author:
        type: string
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
    type: object
  github_com_Firakef1_settle_backend_internal_requests_dto.CreateRequestDTO:
    properties:
      amount:
        type: number
      purpose:
        type: string
      type:
        type: string
      urgency:
        type: string
    type: object
  github_com_Firakef1_settle_backend_internal_requests_dto.PaginationMeta:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_Firakef1_settle_backend_internal_requests_dto.ReceiptResponse:
    properties:
      created_at:
        type: string
      file_path:
        type: string
      id:
        type: string
      ocr_results:
        type: string
      ocr_status:
        type: string
    type: object
  github_com_Firakef1_settle_backend_internal_requests_dto.RequestListItem:
    properties:
      amount:
        type: number
      created_at:
        type: string
      days_pending:
        type: integer
      id:
        type: string
      purpose:
        type: string
      requester:
        $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.RequesterResponse'
      status:
        type: string
      urgency:
        type: string
    type: object
  github_com_Firakef1_settle_backend_internal_requests_dto.RequestListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.RequestListItem'
        type: array
      meta:
        $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.PaginationMeta'
    type: object
  github_com_Firakef1_settle_backend_internal_requests_dto.RequestResponse:
    properties:
      amount:
        type: number
      approval:
        $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.ApprovalResponse'
      comments:
        items:
          $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.CommentResponse'
        type: array
      created_at:
        type: string
      id:
        type: string
      purpose:
        type: string
      receipts:
        items:
          $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.ReceiptResponse'
        type: array
      requester:
        $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.RequesterResponse'
      status:
        type: string
      submitted_at:
        type: string
      timeline:
        items:
          $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.TimelineEvent'
        type: array
      type:
        type: string
      updated_at:
        type: string
      urgency:
        type: string
    type: object
  github_com_Firakef1_settle_backend_internal_requests_dto.RequesterResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  github_com_Firakef1_settle_backend_internal_requests_dto.ResubmitRequestDTO:
    properties:
      amount:
        type: number
      purpose:
        type: string
      receipt_mode:
        type: string
      urgency:
        type: string
    type: object
  github_com_Firakef1_settle_backend_internal_requests_dto.TimelineEvent:
    properties:
      actor:
        type: string
      details:
        type: string
      timestamp:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
  description: Settle backend REST API.
  title: Settle API
  version: "1.0"
paths:
...
  /requests:
    get:
      description: List requests with filtering and pagination
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by urgency
        in: query
        name: urgency
        type: string
      - description: Filter by requester
        in: query
        name: requester_id
        type: string
      - description: Sort by field (e.g., created_at)
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: sort_order
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.RequestListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List requests
      tags:
      - requests
    post:
      consumes:
      - application/json
      description: Create a new payout request
      parameters:
      - description: Create Request DTO
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.CreateRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.RequestResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create request
      tags:
      - requests
  /requests/{id}:
    get:
      description: Get full request details including receipts and comments
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.RequestResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get request details
      tags:
      - requests
  /requests/{id}/comments:
    post:
      consumes:
      - application/json
      description: Add a comment to a request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Add Comment DTO
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.AddCommentDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.CommentResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add comment
      tags:
      - requests
  /requests/{id}/receipts:
    post:
      consumes:
      - multipart/form-data
      description: Upload a receipt to a draft request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Receipt file
        in: formData
        name: receipt
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.ReceiptResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload receipt
      tags:
      - requests
  /requests/{id}/resubmit:
    post:
      consumes:
      - application/json
      description: Resubmit a rejected or failed payout request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Resubmit Request DTO
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.ResubmitRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.RequestResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Resubmit request
      tags:
      - requests
  /requests/{id}/submit:
    post:
      description: Submit a draft payout request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/github_com_Firakef1_settle_backend_internal_requests_dto.RequestResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit request
      tags:
      - requests
  /requests/{id}/withdraw:
    put:
      description: Withdraw a pending or draft payout request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Withdraw request
      tags:
      - requests
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your Bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"